extract ScanPreviousVersions
